requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.2
lightfm==1.17
python-dotenv==1.0.0
gunicorn==21.2.0
scikit-learn==1.3.0
requests==2.31.0
azure-storage-blob==12.17.0
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
azure-monitor-opentelemetry==1.0.0